agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "playwright-agent"
version = "0.1.0"
description = "Playwright agent project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "playwright",
  "pydantic>=2",
  "typer",
  "rich",
  "httpx",
  "tiktoken",
  "anthropic",
  "openai",
  "pyyaml",
  "aiosqlite",
  "structlog",
  "python-dotenv",
]

[project.scripts]
agent = "agent.cli.__main__:main"

[dependency-groups]
dev = [
  "ruff",
  "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
